Map review rating as INT for Hibernate validate
Flyway V8 used TINYINT while the Review entity maps rating as int. Hibernate ddl-auto validate then rejected the schema. Store rating as INT and keep the 1-5 check constraint.

+++ src/main/resources/db/migration/V8__create_reviews.sql
@@ -2,7 +2,7 @@ CREATE TABLE reviews (
     id BIGINT NOT NULL AUTO_INCREMENT,
     product_id BINARY(16) NOT NULL,
     user_id BINARY(16) NOT NULL,
-    rating TINYINT NOT NULL,
+    rating INT NOT NULL,
     comment TEXT NULL,
     created_at DATETIME(6) NOT NULL DEFAULT CURRENT_TIMESTAMP(6),
     CONSTRAINT pk_reviews PRIMARY KEY (id),
